services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator
python-jose[cryptography]
python-multipart
psycopg2-binary

